split sector
