+++ pallets/storage-provider/src/storage_provider.rs
@@ -7,7 +7,10 @@ use frame_support::{
     pallet_prelude::{ConstU32, RuntimeDebug},
     sp_runtime::{BoundedBTreeMap, BoundedVec},
 };
-use primitives::proofs::{RegisteredPoStProof, SectorNumber, SectorSize};
+use primitives::{
+    proofs::RegisteredPoStProof,
+    sector::{SectorNumber, SectorSize},
+};
 use scale_info::TypeInfo;
 use sp_arithmetic::{traits::BaseArithmetic, ArithmeticError};
