+++ primitives/src/proofs/traits.rs
@@ -4,10 +4,8 @@ use cid::Cid;
 use sp_core::ConstU32;
 use sp_runtime::{BoundedBTreeMap, BoundedVec, DispatchError, DispatchResult, RuntimeDebug};
 
-use super::types::{
-    DealId, ProverId, RegisteredPoStProof, RegisteredSealProof, SectorNumber, Ticket,
-};
-use crate::commitment::RawCommitment;
+use super::types::{ProverId, RegisteredPoStProof, RegisteredSealProof, Ticket};
+use crate::{commitment::RawCommitment, sector::SectorNumber};
 
 /// Size of a CID with a 512-bit multihash — i.e. the size of CommR/CommD/CommP
 pub const CID_SIZE_IN_BYTES: u32 = 64;
@@ -51,46 +49,6 @@ pub const MAX_SEAL_PROOF_BYTES: u32 = 1_920;
 /// * <https://github.com/filecoin-project/ref-fvm/blob/32583cc05aa422c8e1e7ba81d56a888ac9d90e61/shared/src/sector/registered_proof.rs#L159>
 pub const MAX_POST_PROOF_BYTES: u32 = 192;
 
-/// Represents functions that are provided by the Market Provider Pallet
-pub trait Market<AccountId, BlockNumber> {
-    /// Verifies a given set of storage deals is valid for sectors being PreCommitted.
-    /// Computes UnsealedCID (CommD) for each sector or None for Committed Capacity sectors.
-    fn verify_deals_for_activation(
-        storage_provider: &AccountId,
-        sector_deals: BoundedVec<SectorDeal<BlockNumber>, ConstU32<MAX_SECTORS_PER_CALL>>,
-    ) -> Result<BoundedVec<Option<Cid>, ConstU32<MAX_SECTORS_PER_CALL>>, DispatchError>;
-
-    /// Activate a set of deals grouped by sector, returning the size and
-    /// extra info about verified deals.
-    /// Sectors' deals are activated in parameter-defined order.
-    /// Each sector's deals are activated or fail as a group, but independently of other sectors.
-    /// Note that confirming all deals fit within a sector is the caller's responsibility
-    /// (and is implied by confirming the sector's data commitment is derived from the deal pieces).
-    fn activate_deals(
-        storage_provider: &AccountId,
-        sector_deals: BoundedVec<SectorDeal<BlockNumber>, ConstU32<MAX_SECTORS_PER_CALL>>,
-        compute_cid: bool,
-    ) -> Result<BoundedVec<ActiveSector<AccountId>, ConstU32<MAX_SECTORS_PER_CALL>>, DispatchError>;
-
-    /// Terminate a set of deals in response to their sector being terminated.
-    ///
-    /// Slashes the provider collateral, refunds the partial unpaid escrow amount to the client.
-    ///
-    /// A sector can be terminated voluntarily — the storage provider terminates the sector —
-    /// or involuntarily — the sector has been faulty for more than 42 consecutive days.
-    ///
-    /// Source: <https://github.com/filecoin-project/builtin-actors/blob/54236ae89880bf4aa89b0dba6d9060c3fd2aacee/actors/market/src/lib.rs#L786-L876>
-    fn on_sectors_terminate(
-        storage_provider: &AccountId,
-        sectors: BoundedVec<SectorNumber, ConstU32<MAX_DEALS_PER_SECTOR>>,
-    ) -> DispatchResult;
-}
-
-pub trait StorageProviderValidation<AccountId> {
-    /// Checks that the storage provider is registered.
-    fn is_registered_storage_provider(storage_provider: &AccountId) -> bool;
-}
-
 /// The minimal information required about a replica, in order to be able to verify
 /// a PoSt over it.
 #[derive(Clone, core::fmt::Debug, PartialEq, Eq)]
@@ -118,31 +76,6 @@ pub trait ProofVerification {
         replicas: BoundedBTreeMap<SectorNumber, PublicReplicaInfo, ConstU32<MAX_SECTORS_PER_PROOF>>,
         proof: BoundedVec<u8, ConstU32<MAX_POST_PROOF_BYTES>>,
     ) -> DispatchResult;
-}
-
-/// Represents functions that are provided by the Randomness Pallet
-pub trait Randomness<BlockNumber> {
-    fn get_randomness(block_number: BlockNumber) -> Result<[u8; 32], DispatchError>;
-}
-
-/// Binds given Sector with the Deals that it should contain
-/// It's used as a data transfer object for extrinsics `verify_deals_for_activation`
-/// as well as `activate deals`.
-/// It represents a sector that should be activated and it's deals.
-#[derive(RuntimeDebug)]
-pub struct SectorDeal<BlockNumber> {
-    /// Number of the sector that is supposed to contain the deals
-    pub sector_number: SectorNumber,
-    /// Time when the sector expires.
-    /// If sector expires before some of the deals end, than it's violation and sector is rejected.
-    pub sector_expiry: BlockNumber,
-    /// Used to extract the size of a sector
-    /// All of the deals must fit within the seal proof's sector size.
-    /// If not, sector is rejected.
-    pub sector_type: RegisteredSealProof,
-    /// Deals Ids that are supposed to be activated.
-    /// If any of those is invalid, whole activation is rejected.
-    pub deal_ids: BoundedVec<DealId, ConstU32<MAX_DEALS_PER_SECTOR>>,
 }
 
 /// A sector with all of its active deals.
