+++ storage-provider/server/src/pipeline/mod.rs
@@ -9,8 +9,9 @@ use polka_storage_proofs::porep::{
 use polka_storage_provider_common::rpc::ServerInfo;
 use primitives::{
     commitment::{CommD, CommP, CommR, Commitment},
-    proofs::{derive_prover_id, SectorNumber},
+    proofs::derive_prover_id,
     randomness::{draw_randomness, DomainSeparationTag},
+    sector::SectorNumber,
 };
 use storagext::{
     types::{
