+++ storagext/cli/src/cmd/storage_provider.rs
@@ -1,7 +1,7 @@
 use std::time::Duration;
 
 use clap::Subcommand;
-use primitives::proofs::{RegisteredPoStProof, SectorNumber};
+use primitives::{proofs::RegisteredPoStProof, sector::SectorNumber};
 use storagext::{
     deser::DeserializablePath,
     multipair::MultiPairSigner,
